Bid price is not accepting decimal values. For orders with extra items added, only decimal places bid amount are getting accepted.

+++ database/migrations/2021_09_28_174731_alter_add_distance_column_organization.php
@@ -14,7 +14,7 @@ class AlterAddDistanceColumnOrganization extends Migration
     public function up()
     {
         Schema::table('organizations', function (Blueprint $table) {
-            $table->tinyInteger('base_distance')->nullable()->after('zone_id');
+            $table->decimal('base_distance', 10, 2)->nullable()->after('zone_id');
         });
     }
